backend/tision_backend/requirements.txt
asgiref
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv